+++ src/main/java/plugily/projects/buildbattle/arena/impl/SoloArena.java
@@ -371,7 +371,7 @@ public class SoloArena extends BaseArena {
 
   @Override
   public void updateBossBar() {
-    if (!getPlugin().getConfigPreferences().getOption(ConfigPreferences.Option.BOSSBAR_ENABLED)) {
+    if (getGameBar() == null) {
       return;
     }
     switch (getArenaState()) {